+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 [Full diff](https://github.com/sider/runners/compare/0.25.2...HEAD)
 
+- Use git-ls-files instead of git-check-ignore [#1148](https://github.com/sider/runners/pull/1148)
+
 ## 0.25.2
 
 [Full diff](https://github.com/sider/runners/compare/0.25.1...0.25.2)